requirement.txt
amqp==2.2.2
attrs==17.4.0
Babel==2.6.0
billiard==3.5.0.3
boto==2.49.0
boto3==1.9.61
botocore==1.12.61
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
cookies==2.2.1
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.15
django-log-request-id==1.3.2
django-phonenumber-field==2.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
drf-yasg==1.11.1
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
MarkupSafe==1.0
model-mommy==1.5.1
more-itertools==4.1.0
openapi-codec==1.3.2
pbr==4.0.1
phonenumbers==8.9.16
pika==0.11.2
pkg-resources==0.0.0
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
Pygments==2.2.0
PyJWT==1.6.1
PySocks==1.6.8
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
responses==0.9.0
ruamel.yaml==0.15.81
s3transfer==0.1.13
simplejson==3.13.2
six==1.11.0
text-unidecode==1.2
twilio==6.21.0
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.17
vine==1.1.4